pyproject.toml
[tool.poetry]
name = "tasks-ptb"
version = "0.1.0"
description = ""
authors = ["Vasilevich Dmitry <vasilevich.dmt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
python-decouple = "^3.8"
aiogram = "^3.13.1"
aiosqlite = "^0.20.0"
python-telegram-bot = "^21.6"
black = "^24.10.0"
alembic = "^1.13.3"
isort = "^5.13.2"
pydantic = "^2.9.2"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
